Alter the core Object class (WIP)
This commit contains a significant amount of reworking around Object
hierarchies, now enforcing a strict top-down approach. As in,
child Objects no longer have any information regarding their parents.
This should simplify some of the processing and, intuitively, match
what we might expect out of cognition.

The decomposition pipeline is still being fixed up, so the branch
is in a (WIP) state. The base-level Object behavior and tests are
up to date, the new Generator class is working with tests.